moves transform message tests into separate test file
